# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready URL shortener API with JWT authentication, refresh-token rotation, and Google OAuth.\n\n## Authentication\nMost protected endpoints require an access token in the Authorization header:\n`Authorization: Bearer \u003caccess_token\u003e`\n\nAccess tokens are short-lived. Refresh tokens are stored in an HttpOnly cookie and used to obtain a new access token.\n\n## URL Expiration\nAuthenticated users receive longer URL expiration and cache durations.\nUnauthenticated URLs expire earlier.\n\n## Rate Limiting\nURL endpoints are rate limited to protect the service from abuse.",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users/login": {
            "post": {
                "description": "Authenticates a user with email and password.\n\nReturns a short-lived access token in the response body.\nA refresh token is stored in an HttpOnly Secure cookie.\n\nFor security, invalid email and invalid password return the same error.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Clears the refresh token cookie from the current browser.\n\nThis endpoint does not invalidate all tokens server-side.\nUse Logout All Sessions to invalidate tokens across devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Logout current device",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/logout-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates all active access and refresh tokens for the user.\n\nThe user's token version is incremented, causing previously issued\naccess and refresh tokens to become invalid.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Logout all sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LogoutAllResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/refresh": {
            "post": {
                "description": "Issues a new access token using the refresh_token HttpOnly cookie.\n\nThe refresh token is rotated after a successful refresh.\nThe new refresh token is returned as a replacement HttpOnly Secure cookie.\n\nClients should call this endpoint when the access token expires.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Refresh access token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.RefreshResponse"
                        }
                    },
                    "401": {
                        "description": "Missing, expired, or invalid refresh token",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "description": "Registers a new user account.\n\nFor security, this endpoint returns the same successful response\nwhether the email is newly registered or already exists.\nThis prevents account enumeration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user account",
                "parameters": [
                    {
                        "description": "Signup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.SignupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.Error"
                },
                "metaData": {
                    "$ref": "#/definitions/github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.MetaData"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_response_envelope.MetaData": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LogoutAllResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.RefreshResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.SignupRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "description": "bcrypt silently truncates/errors past 72 bytes, so cap it here.",
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "github_com_tim8912097887-sys_url-shortener_internal_shared_schema_user_schema.SignupResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your access token in the format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshCookie": {
            "description": "HttpOnly refresh token cookie automatically sent by the browser.",
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    }
}